[REF] Add ASSERT_EQ_OS_ERR macro
Macros verifying error codes do not scale well, as every time new error
is introduced, `Test.h` must be adapted.

It is better to instead offer a macro that is dedicated to verifying
any error code.

This commit is a first step of the bigger refactoring i.e. getting rid
of the error code verification macros.